Add files via upload
Calculating RMSD in relaxed atomic positions for .traj files, displaying in plots. Calling relaxed slabs to be used in surface energy calculations for pure or alloyed slabs, then displaying the error in these calculations in plots. Code included for optimizing lattice constants and assigning to new atom constructions for use in the calculations